Move FromStr to core::str
