+++ src/libregex/re.rs
@@ -10,7 +10,6 @@
 
 use std::collections::HashMap;
 use std::fmt;
-use std::from_str::from_str;
 use std::str::{MaybeOwned, Owned, Slice};
 
 use compile::Program;
