+++ src/libtest/lib.rs
@@ -53,13 +53,13 @@ use std::cmp;
 use std::f64;
 use std::fmt::Show;
 use std::fmt;
-use std::from_str::FromStr;
 use std::io::fs::PathExtensions;
 use std::io::stdio::StdWriter;
 use std::io::{File, ChanReader, ChanWriter};
 use std::io;
 use std::num::{Float, FloatMath, Int};
 use std::os;
+use std::str::FromStr;
 use std::string::String;
 use std::task::TaskBuilder;
 use std::time::Duration;
